fix: fixed project structure and make sure pass is always hashed
